+++ openai-client/src/commonMain/kotlin/com.aallam.openai.client/internal/http/HttpRequester.kt
@@ -1,20 +1,31 @@
 package com.aallam.openai.client.internal.http
 
-import com.aallam.openai.client.Closeable
+import com.aallam.openai.api.run.AssistantStreamEvent
 import io.ktor.client.*
+import io.ktor.client.plugins.sse.ClientSSESession
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
+import io.ktor.sse.ServerSentEvent
 import io.ktor.util.reflect.*
+import kotlinx.coroutines.flow.Flow
 
 /**
  * Http request performer.
  */
-internal interface HttpRequester : Closeable {
+internal interface HttpRequester : AutoCloseable {
 
     /**
      * Perform an HTTP request and get a result.
      */
     suspend fun <T : Any> perform(info: TypeInfo, block: suspend (HttpClient) -> HttpResponse): T
+
+    /**
+     * Perform an HTTP request and process emitted server-side events.
+     *
+     */
+    suspend fun performSse(
+        builderBlock: HttpRequestBuilder.() -> Unit
+    ): Flow<AssistantStreamEvent>
 
     /**
      * Perform an HTTP request and get a result.