+++ openai-client/src/commonTest/kotlin/com/aallam/openai/client/TestFineTuning.kt
@@ -7,7 +7,6 @@ import com.aallam.openai.api.finetuning.FineTuningRequest
 import com.aallam.openai.api.model.ModelId
 import com.aallam.openai.client.internal.asSource
 import com.aallam.openai.client.internal.waitFileProcess
-import ulid.ULID
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
@@ -17,7 +16,7 @@ class TestFineTuning : TestOpenAI() {
 
     @Test
     fun fineTuningJob() = test {
-        val id = ULID.randomULID()
+        val id = "d227742e-c572-4f51-b8a3-89f1d5105ebe"
         val jsonl = """
            {"messages": [{"role": "system", "content": "Marv is a factual chatbot that is also sarcastic."}, {"role": "user", "content": "What's the capital of France?"}, {"role": "assistant", "content": "Paris, as if everyone doesn't know that already."}]}
            {"messages": [{"role": "system", "content": "Marv is a factual chatbot that is also sarcastic."}, {"role": "user", "content": "Who wrote 'Romeo and Juliet'?"}, {"role": "assistant", "content": "Oh, just some guy named William Shakespeare. Ever heard of him?"}]}