+++ openai-client/src/jvmMain/kotlin/com/aallam/openai/client/Closeable.kt
@@ -1,3 +1,0 @@
-package com.aallam.openai.client
-
-public actual typealias Closeable = AutoCloseable