+++ sample/jvm/src/main/kotlin/com/aallam/openai/sample/jvm/Whisper.kt
@@ -5,13 +5,11 @@ import com.aallam.openai.api.audio.TranslationRequest
 import com.aallam.openai.api.file.FileSource
 import com.aallam.openai.api.model.ModelId
 import com.aallam.openai.client.OpenAI
-import okio.FileSystem
-import okio.Path.Companion.toPath
 
 suspend fun whisper(openAI: OpenAI) {
     println("\n>️ Create transcription...")
     val transcriptionRequest = TranscriptionRequest(
-        audio = FileSource(path = "micro-machines.wav".toPath(), fileSystem = FileSystem.RESOURCES),
+        audio = FileSource(path = Resources.path("micro-machines.wav")),
         model = ModelId("whisper-1"),
     )
     val transcription = openAI.transcription(transcriptionRequest)
@@ -19,7 +17,7 @@ suspend fun whisper(openAI: OpenAI) {
 
     println("\n>️ Create translation...")
     val translationRequest = TranslationRequest(
-        audio = FileSource(path = "multilingual.wav".toPath(), fileSystem = FileSystem.RESOURCES),
+        audio = FileSource(path = Resources.path("multilingual.wav")),
         model = ModelId("whisper-1"),
     )
     val translation = openAI.translation(translationRequest)