+++ sample/jvm/src/main/kotlin/com/aallam/openai/sample/jvm/images.kt
@@ -5,8 +5,6 @@ import com.aallam.openai.api.image.ImageCreation
 import com.aallam.openai.api.image.ImageEdit
 import com.aallam.openai.api.image.ImageSize
 import com.aallam.openai.client.OpenAI
-import okio.FileSystem
-import okio.Path.Companion.toPath
 
 suspend fun images(openAI: OpenAI) {
     println("\n> Create images...")
@@ -21,8 +19,8 @@ suspend fun images(openAI: OpenAI) {
 
     println("\n> Edit images...")
     val imageEdit = ImageEdit(
-        image = FileSource(path = "image.png".toPath(), fileSystem = FileSystem.RESOURCES),
-        mask = FileSource(path = "image.png".toPath(), fileSystem = FileSystem.RESOURCES),
+        image = FileSource(path = Resources.path("image.png")),
+        mask = FileSource(path = Resources.path("image.png")),
         prompt = "a sunlit indoor lounge area with a pool containing a flamingo",
         n = 1,
         size = ImageSize.is1024x1024,